[MOB-4300] Move Omnibox related code for BrowserViewController into dedicated file
